+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [1.0.17](https://github.com/david-04/launchpad/releases/tag/v1.0.17) (unreleased)
 
 - Include CSS files when formatting with Biome via Makefile
+- Remove the `$schema` property from `biome.json`
 
 ## [1.0.16](https://github.com/david-04/launchpad/releases/tag/v1.0.16) (2025-07-05)